Updates to Online Portfolio
Lots of changes to update my online portfolio for 2019. Biggest things are: 1) resume now redirects to my LinkedIn profile page (everywhere), and 2) I combined the Software Development and Website pages into the single SD page. This allows me to have 10 decent projects with no "coming soons". 3) I updated the videos show on my media page to more current material, and 4) the demo reel is no longer on the home page, just the media page, and it's in the header area.